Refactored out the Simplex optimization plotting from Worker
Partially addressed #24 by refactoring the simplex plotting into a mixin class. Now I just need to add this functionality to the `WorkerFamily` class and test it.
